pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "holler"
version = "0.1.0"
description = "Push-to-talk voice dictation for Linux, powered by Groq Whisper + LLM correction"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Nicolas Jesse" }]
keywords = ["voice", "dictation", "speech-to-text", "whisper", "groq", "linux", "wayland", "cosmic"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: X11 Applications :: Qt",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]
dependencies = [
    "PyQt5>=5.15.0",
    "sounddevice>=0.4.6",
    "numpy>=1.24.0",
    "groq>=0.4.0",
    "openai>=1.0.0",
    "pynput>=1.7.6",
    "evdev>=1.6.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.5.0",
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-mock>=3.12",
]

[project.scripts]
holler = "src.main:main"

[project.urls]
Repository = "https://github.com/nicolasjesse/holler"
Issues = "https://github.com/nicolasjesse/holler/issues"

[tool.setuptools.packages.find]
include = ["src*"]

# --- Ruff ---

[tool.ruff]
line-length = 100
target-version = "py310"
extend-exclude = [".env", "build", "dist"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # bugbear
    "UP",  # pyupgrade
    "SIM", # flake8-simplify
    "RUF", # ruff-specific
]
ignore = [
    "E501",  # line too long - handled by formatter
    "B008",  # function call in default argument (fine for FastAPI-style code)
    "RUF001", "RUF002", "RUF003",  # ambiguous unicode (× is intentional in math docs)
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["F401", "F811"]  # unused imports/redefs ok in tests
"**/__init__.py" = ["F401"]    # re-exports are idiomatic
"src/input/hotkey_evdev.py" = ["F401"]  # evdev probe imports check availability

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# --- Pytest ---

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
]
markers = [
    "integration: requires live network / API keys",
    "gui: requires display server (PyQt)",
    "audio: requires microphone / sounddevice",
]
